refactor/rep: move UnifiedReputationChange to separate file
